docs(README.md): remove CMake prerequisite and update build and run instructions for the project
The CMake prerequisite has been removed as it is no longer required for building the project. The build instructions have been updated to use `platformio run` command in the server directory to compile the project. The run instructions have been updated to use `platformio run --target upload` command to upload the compiled code to the Olimax ESP32-EVB board.